Don't change focus due to mouse movement
A longstanding irritant to me has been if I'm editing a value (such
as the model X position) and I move the mouse over the platter, focus
is lost in the input box, and digits and tab behave in unexpected ways.

This is due to the canvas taking focus on mouse movement apparently so
that keyboard shortcuts for the items in the left bar work.  Since I've
never tried to use those keyboard shortcuts, it's more important to me
that the input box I clicked on receives the input I try to enter than
being able to use keyboard shortcuts for something else.

squashed with:
* Add new option to disable focus-follows-mouse for platter
* Fix egregious copy pasta.