format/AnycubicSLA: refactor Anycubic SLA format and add Photon Mono and Mono SE printers  (#9929)
* format/sla: Rename pwmx format to AnycubicSLA

The Anycubic Photon SLA printer familiy is shipped with the PhotonWorkshop slicer.
This slicer generates the sliced archives. These archives have per-printer
extensions, not only pwmx. The name is -most of the times- comprised of the
PhontonWorkshop "pw" bit and the model name. "mo" for the Photon Mono, "mx"
for the Mono X.

Therefore the format name "pwmx" is incorrect and we are renaming it to AnycubicSLA.
On top of it we are introducing a helper macro to connect file extension and printer
definition as most printers use extremely similiar formats.

Signed-off-by: Mimoja <git@mimoja.de>

* format/AnycubicSLA: Add missing fields

The AnycubicSLA format description is derived from reverse engineering of
the PhotonWorkshop output files.

While the initial Photon devices had their binary headers with version 1
we have seen newer models with additional versions. Namely 515, 516 and 517.
We are adding the version handling to the AnycubicSLA exporter to prepare
for future version handling.
Some fields were missing for Version 1 which are added.

Signed-off-by: Mimoja <git@mimoja.de>

* PrintConfig/sla: Move material notes to simple view

As the Anycubic Photon Mono X uses the material notes to configure the
printers parameters we need to change the visibility to allow every user acces.

This will change the default behaviour for non Anycubic SLA printers.

Signed-off-by: Mimoja <git@mimoja.de>

* format/AnycubicSLA: Expose Antialiasing via material notes

Similiar to how the other machine configurations are exposed via the material notes
we are adding the Antialiasing config.

Signed-off-by: Mimoja <git@mimoja.de>

* Printer/sla: Add Anycubic Photon Mono and Mono SE

The Photon Mono and Mono SE are format compatible printers with the Mono X.
They support Version 1 and Version 515 binary archives.
We are implementing them as Version 1 priters to reduce the overhead and keep the code
in line with the Mono X.

The addition as Version 1 printers leaves some features unexposed, most noteably the Antialiasing
level configuration which is now always the maximum.

Given that the printers check the eligability of sliced files by file extension we are
poised to create per-printer sla_print default configurations to overwrite output_filename_format.

Tested: Successfully printed multiple objects. Changing the layer parameter on the onscreen display
succeeded.

---------

Signed-off-by: Mimoja <git@mimoja.de>